Part 6 Exc 6.7 anecdotes, step 5
